requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://docs.astronomer.io/astro/runtime-image-architecture#provider-packages
# Astro Runtime includes the following pre-installed providers packages: https://docs.astronomer.io/astro/runtime-image-architecture#provider-packages
# apache-airflow-providers-mysql==5.1.0
apache-airflow-providers-microsoft-mssql
astronomer-cosmos
apache-airflow-providers-apache-spark
pyspark==3.5.0
apache-airflow-providers-trino